Split out loader utilities into a discovery module
Some methods will be useful for a resolver based implementation,
in addition to the loader.  Let's split it out of the same class
into a common utility.

This will also help so that, in the future, the loader based
implementation could be removed more easily.
